feat(core): separate session routes and UTs
