+++ packages/core/src/routes/session.test.ts
@@ -1,18 +1,7 @@
-/* eslint-disable max-lines */
 import { User } from '@logto/schemas';
 import { Provider } from 'oidc-provider';
 
-import {
-  mockUser,
-  mockAliyunDmConnectorInstance,
-  mockAliyunSmsConnectorInstance,
-  mockFacebookConnectorInstance,
-  mockGithubConnectorInstance,
-  mockGoogleConnectorInstance,
-  mockWechatConnectorInstance,
-  mockWechatNativeConnectorInstance,
-} from '@/__mocks__';
-import { ConnectorType } from '@/connectors/types';
+import { mockUser } from '@/__mocks__';
 import RequestError from '@/errors/RequestError';
 import { createRequester } from '@/utils/test-utils';
 
@@ -37,29 +26,6 @@ jest.mock('@/lib/user', () => ({
   }),
   updateLastSignInAt: async (...args: unknown[]) => updateUserById(...args),
 }));
-jest.mock('@/lib/social', () => ({
-  ...jest.requireActual('@/lib/social'),
-  async findSocialRelatedUser() {
-    return ['phone', { id: 'user1', identities: {} }];
-  },
-  async getUserInfoByAuthCode(connectorId: string, data: { code: string }) {
-    if (connectorId === '_connectorId') {
-      throw new RequestError({
-        code: 'session.invalid_connector_id',
-        status: 422,
-        connectorId,
-      });
-    }
-
-    if (data.code === '123456') {
-      return { id: 'id' };
-    }
-
-    // This mocks the case that can not get userInfo with access token and auth code
-    // (most likely third-party social connectors' problem).
-    throw new Error(' ');
-  },
-}));
 const insertUser = jest.fn(async (..._args: unknown[]) => ({ id: 'id' }));
 const findUserById = jest.fn(async (): Promise<User> => mockUser);
 const updateUserById = jest.fn(async (..._args: unknown[]) => ({ id: 'id' }));
@@ -75,59 +41,6 @@ jest.mock('@/queries/user', () => ({
     connectorId === 'connectorId' && userId === 'id',
   hasUserWithPhone: async (phone: string) => phone === '13000000000',
   hasUserWithEmail: async (email: string) => email === 'a@a.com',
-}));
-const sendPasscode = jest.fn(async () => ({ connector: { id: 'connectorIdValue' } }));
-jest.mock('@/lib/passcode', () => ({
-  createPasscode: async () => ({ id: 'id' }),
-  sendPasscode: async () => sendPasscode(),
-  verifyPasscode: async (_a: unknown, _b: unknown, code: string) => {
-    if (code !== '1234') {
-      throw new RequestError('passcode.code_mismatch');
-    }
-  },
-}));
-const getAuthorizationUri = jest.fn(async (_state: string, _redirectUri: string) => {
-  return '';
-});
-const getConnectorInstanceById = jest.fn(async (connectorId: string) => {
-  const connector = {
-    enabled: connectorId === 'social_enabled',
-  };
-  const metadata = {
-    id:
-      connectorId === 'social_enabled'
-        ? 'social_enabled'
-        : connectorId === 'social_disabled'
-        ? 'social_disabled'
-        : 'others',
-    type: connectorId.startsWith('social') ? ConnectorType.Social : ConnectorType.SMS,
-  };
-
-  return { connector, metadata, getAuthorizationUri };
-});
-const getConnectorInstances = jest.fn(async () => [
-  mockAliyunDmConnectorInstance,
-  mockAliyunSmsConnectorInstance,
-  mockFacebookConnectorInstance,
-  mockGithubConnectorInstance,
-  mockGoogleConnectorInstance,
-  mockWechatConnectorInstance,
-  mockWechatNativeConnectorInstance,
-]);
-jest.mock('@/connectors', () => ({
-  getSocialConnectorInstanceById: async (connectorId: string) => {
-    const connectorInstance = await getConnectorInstanceById(connectorId);
-
-    if (connectorInstance.metadata.type !== ConnectorType.Social) {
-      throw new RequestError({
-        code: 'entity.not_found',
-        status: 404,
-      });
-    }
-
-    return connectorInstance;
-  },
-  getConnectorInstances: async () => getConnectorInstances(),
 }));
 
 const grantSave = jest.fn(async () => 'finalGrantId');
@@ -234,277 +147,6 @@ describe('sessionRoutes', () => {
     });
   });
 
-  describe('POST /session/sign-in/passwordless/sms/send-passcode', () => {
-    beforeAll(() => {
-      interactionDetails.mockResolvedValueOnce({
-        jti: 'jti',
-      });
-    });
-    it('it should call sendPasscode', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/sms/send-passcode')
-        .send({ phone: '13000000000' });
-      expect(response.statusCode).toEqual(204);
-      expect(sendPasscode).toHaveBeenCalled();
-    });
-    it('throw error if phone does not exist', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/sms/send-passcode')
-        .send({ phone: '13000000001' });
-      expect(response.statusCode).toEqual(422);
-    });
-  });
-
-  describe('POST /session/sign-in/passwordless/sms/verify-passcode', () => {
-    it('assign result and redirect', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/sms/verify-passcode')
-        .send({ phone: '13000000000', code: '1234' });
-      expect(response.statusCode).toEqual(200);
-      expect(response.body).toHaveProperty('redirectTo');
-      expect(interactionResult).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.anything(),
-        expect.objectContaining({ login: { accountId: 'id' } }),
-        expect.anything()
-      );
-    });
-    it('throw error if phone does not exist', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/sms/verify-passcode')
-        .send({ phone: '13000000001', code: '1234' });
-      expect(response.statusCode).toEqual(422);
-    });
-    it('throw error if verifyPasscode failed', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/sms/verify-passcode')
-        .send({ phone: '13000000000', code: '1231' });
-      expect(response.statusCode).toEqual(400);
-    });
-  });
-
-  describe('POST /session/sign-in/passwordless/email/send-passcode', () => {
-    beforeAll(() => {
-      interactionDetails.mockResolvedValue({
-        jti: 'jti',
-      });
-    });
-    it('it should call sendPasscode', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/email/send-passcode')
-        .send({ email: 'a@a.com' });
-      expect(response.statusCode).toEqual(204);
-      expect(sendPasscode).toHaveBeenCalled();
-    });
-    it('throw error if email does not exist', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/email/send-passcode')
-        .send({ email: 'b@a.com' });
-      expect(response.statusCode).toEqual(422);
-    });
-  });
-
-  describe('POST /session/sign-in/passwordless/email/verify-passcode', () => {
-    it('assign result and redirect', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/email/verify-passcode')
-        .send({ email: 'a@a.com', code: '1234' });
-      expect(response.statusCode).toEqual(200);
-      expect(response.body).toHaveProperty('redirectTo');
-      expect(interactionResult).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.anything(),
-        expect.objectContaining({ login: { accountId: 'id' } }),
-        expect.anything()
-      );
-    });
-    it('throw error if email does not exist', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/email/send-passcode')
-        .send({ email: 'b@a.com' });
-      expect(response.statusCode).toEqual(422);
-    });
-    it('throw error if verifyPasscode failed', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/email/verify-passcode')
-        .send({ email: 'a@a.com', code: '1231' });
-      expect(response.statusCode).toEqual(400);
-    });
-  });
-
-  describe('POST /session/sign-in/social', () => {
-    it('should throw when redirectURI is invalid', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social').send({
-        connectorId: 'social_enabled',
-        state: 'state',
-        redirectUri: 'logto.dev',
-      });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('sign-in with social and redirect', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social').send({
-        connectorId: 'social_enabled',
-        state: 'state',
-        redirectUri: 'https://logto.dev',
-      });
-      expect(response.body).toHaveProperty('redirectTo', '');
-    });
-
-    it('throw error when sign-in with social but miss state', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social').send({
-        connectorId: 'social_enabled',
-        redirectUri: 'https://logto.dev',
-      });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error when sign-in with social but miss redirectUri', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social').send({
-        connectorId: 'social_enabled',
-        state: 'state',
-      });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error when connector is disabled', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social').send({
-        connectorId: 'social_disabled',
-        state: 'state',
-        redirectUri: 'https://logto.dev',
-      });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error when no social connector is found', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social').send({
-        connectorId: 'others',
-        state: 'state',
-        redirectUri: 'https://logto.dev',
-      });
-      expect(response.statusCode).toEqual(404);
-    });
-  });
-
-  describe('POST /session/sign-in/social/auth', () => {
-    it('throw error when auth code is wrong', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social/auth').send({
-        connectorId: 'connectorId',
-        state: 'state',
-        redirectUri: 'https://logto.dev',
-        code: '123455',
-      });
-      expect(response.statusCode).toEqual(500);
-    });
-
-    it('throw error when code is provided but connector can not be found', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social/auth').send({
-        connectorId: '_connectorId',
-        state: 'state',
-        redirectUri: 'https://logto.dev',
-        code: '123456',
-      });
-      expect(response.statusCode).toEqual(422);
-    });
-
-    it('get and add user info with auth code, as well as assign result and redirect', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social/auth').send({
-        connectorId: 'connectorId',
-        data: {
-          state: 'state',
-          redirectUri: 'https://logto.dev',
-          code: '123456',
-        },
-      });
-      expect(updateUserById).toHaveBeenCalledWith(
-        'id',
-        expect.objectContaining({
-          identities: { connectorId: { userId: 'id', details: { id: 'id' } } },
-        })
-      );
-      expect(response.body).toHaveProperty('redirectTo');
-      expect(interactionResult).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.anything(),
-        expect.objectContaining({ login: { accountId: 'id' } }),
-        expect.anything()
-      );
-    });
-
-    it('throw error when identity exists', async () => {
-      const response = await sessionRequest.post('/session/sign-in/social/auth').send({
-        connectorId: '_connectorId_',
-        data: {
-          state: 'state',
-          redirectUri: 'https://logto.dev',
-          code: '123456',
-        },
-      });
-      expect(interactionResult).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.anything(),
-        expect.objectContaining({
-          socialUserInfo: { connectorId: '_connectorId_', userInfo: { id: 'id' } },
-        }),
-        expect.anything()
-      );
-      expect(response.statusCode).toEqual(422);
-    });
-  });
-
-  describe('POST /session/sign-in/bind-social-related-user', () => {
-    it('throw if session is not authorized', async () => {
-      await expect(
-        sessionRequest
-          .post('/session/sign-in/bind-social-related-user')
-          .send({ connectorId: 'connectorId' })
-      ).resolves.toHaveProperty('statusCode', 400);
-    });
-    it('throw if no social info in session', async () => {
-      interactionDetails.mockResolvedValueOnce({
-        result: { login: { accountId: 'user1' } },
-      });
-      await expect(
-        sessionRequest
-          .post('/session/sign-in/bind-social-related-user')
-          .send({ connectorId: 'connectorId' })
-      ).resolves.toHaveProperty('statusCode', 400);
-    });
-    it('updates user identities and sign in', async () => {
-      interactionDetails.mockResolvedValueOnce({
-        result: {
-          login: { accountId: 'user1' },
-          socialUserInfo: {
-            connectorId: 'connectorId',
-            userInfo: { id: 'connectorUser', phone: 'phone' },
-          },
-        },
-      });
-      const response = await sessionRequest.post('/session/sign-in/bind-social-related-user').send({
-        connectorId: 'connectorId',
-      });
-      expect(response.statusCode).toEqual(200);
-      expect(updateUserById).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.objectContaining({
-          identities: {
-            connectorId: {
-              details: { id: 'connectorUser', phone: 'phone' },
-              userId: 'connectorUser',
-            },
-          },
-        })
-      );
-      expect(response.body).toHaveProperty('redirectTo');
-      expect(interactionResult).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.anything(),
-        expect.objectContaining({ login: { accountId: 'user1' } }),
-        expect.anything()
-      );
-    });
-  });
-
   describe('POST /session/register/username-password', () => {
     it('assign result and redirect', async () => {
       const response = await sessionRequest
@@ -540,260 +182,6 @@ describe('sessionRoutes', () => {
         .post('/session/register/username-password')
         .send({ username: 'username1', password: 'password' });
       expect(response.statusCode).toEqual(422);
-    });
-  });
-
-  describe('POST /session/register/passwordless/sms/send-passcode', () => {
-    beforeAll(() => {
-      interactionDetails.mockResolvedValueOnce({
-        jti: 'jti',
-      });
-    });
-
-    it('it should call sendPasscode', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/send-passcode')
-        .send({ phone: '13000000001' });
-      expect(response.statusCode).toEqual(204);
-      expect(sendPasscode).toHaveBeenCalled();
-    });
-
-    it('throw error if phone not valid (charactors other than digits)', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/send-passcode')
-        .send({ phone: '1300000000a' });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error if phone not valid (without digits)', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/send-passcode')
-        .send({ phone: 'abcdefg' });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error if phone exists', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/send-passcode')
-        .send({ phone: '13000000000' });
-      expect(response.statusCode).toEqual(422);
-    });
-  });
-
-  describe('POST /session/register/passwordless/sms/verify-passcode', () => {
-    it('assign result and redirect', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/verify-passcode')
-        .send({ phone: '13000000001', code: '1234' });
-      expect(response.statusCode).toEqual(200);
-      expect(response.body).toHaveProperty('redirectTo');
-      expect(insertUser).toHaveBeenCalledWith(
-        expect.objectContaining({ id: 'user1', primaryPhone: '13000000001' })
-      );
-      expect(interactionResult).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.anything(),
-        expect.objectContaining({ login: { accountId: 'user1' } }),
-        expect.anything()
-      );
-    });
-
-    it('throw error if phone not valid (characters other than digits)', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/verify-passcode')
-        .send({ phone: '1300000000a', code: '1234' });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error if phone not valid (without digits)', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/verify-passcode')
-        .send({ phone: 'abcdefg', code: '1234' });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error if phone exists', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/verify-passcode')
-        .send({ phone: '13000000000', code: '1234' });
-      expect(response.statusCode).toEqual(422);
-    });
-
-    it('throw error if verifyPasscode failed', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/sms/verify-passcode')
-        .send({ phone: '13000000001', code: '1231' });
-      expect(response.statusCode).toEqual(400);
-    });
-  });
-
-  describe('POST /session/register/passwordless/email/send-passcode', () => {
-    beforeAll(() => {
-      interactionDetails.mockResolvedValueOnce({
-        jti: 'jti',
-      });
-    });
-
-    it('it should call sendPasscode', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/email/send-passcode')
-        .send({ email: 'b@a.com' });
-      expect(response.statusCode).toEqual(204);
-      expect(sendPasscode).toHaveBeenCalled();
-    });
-
-    it('throw error if email not valid', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/email/send-passcode')
-        .send({ email: 'aaa.com' });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error if email exists', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/email/send-passcode')
-        .send({ email: 'a@a.com' });
-      expect(response.statusCode).toEqual(422);
-    });
-  });
-
-  describe('POST /session/register/passwordless/email/verify-passcode', () => {
-    it('assign result and redirect', async () => {
-      const response = await sessionRequest
-        .post('/session/register/passwordless/email/verify-passcode')
-        .send({ email: 'b@a.com', code: '1234' });
-      expect(response.statusCode).toEqual(200);
-      expect(response.body).toHaveProperty('redirectTo');
-      expect(insertUser).toHaveBeenCalledWith(
-        expect.objectContaining({ id: 'user1', primaryEmail: 'b@a.com' })
-      );
-      expect(interactionResult).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.anything(),
-        expect.objectContaining({ login: { accountId: 'user1' } }),
-        expect.anything()
-      );
-    });
-
-    it('throw error if email not valid', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/email/send-passcode')
-        .send({ email: 'aaa.com' });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error if email exist', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/email/send-passcode')
-        .send({ email: 'b@a.com' });
-      expect(response.statusCode).toEqual(422);
-    });
-
-    it('throw error if verifyPasscode failed', async () => {
-      const response = await sessionRequest
-        .post('/session/sign-in/passwordless/email/verify-passcode')
-        .send({ email: 'a@a.com', code: '1231' });
-      expect(response.statusCode).toEqual(400);
-    });
-  });
-
-  describe('POST /session/register/social', () => {
-    it('register with social, assign result and redirect', async () => {
-      interactionDetails.mockResolvedValueOnce({
-        jti: 'jti',
-        result: {
-          socialUserInfo: { connectorId: 'connectorId', userInfo: { id: 'user1' } },
-        },
-      });
-      const response = await sessionRequest
-        .post('/session/register/social')
-        .send({ connectorId: 'connectorId' });
-      expect(insertUser).toHaveBeenCalledWith(
-        expect.objectContaining({
-          id: 'user1',
-          identities: { connectorId: { userId: 'user1', details: { id: 'user1' } } },
-        })
-      );
-      expect(response.body).toHaveProperty('redirectTo');
-      expect(interactionResult).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.anything(),
-        expect.objectContaining({ login: { accountId: 'user1' } }),
-        expect.anything()
-      );
-    });
-
-    it('throw error if no result can be found in interactionResults', async () => {
-      interactionDetails.mockResolvedValueOnce({});
-      const response = await sessionRequest
-        .post('/session/register/social')
-        .send({ connectorId: 'connectorId' });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error if result parsing fails', async () => {
-      interactionDetails.mockResolvedValueOnce({ result: { login: { accountId: 'id' } } });
-      const response = await sessionRequest
-        .post('/session/register/social')
-        .send({ connectorId: 'connectorId' });
-      expect(response.statusCode).toEqual(400);
-    });
-
-    it('throw error when user with identity exists', async () => {
-      interactionDetails.mockResolvedValueOnce({
-        result: {
-          login: { accountId: 'user1' },
-          socialUserInfo: { connectorId: 'connectorId', userInfo: { id: 'id' } },
-        },
-      });
-      const response = await sessionRequest
-        .post('/session/register/social')
-        .send({ connectorId: 'connectorId' });
-      expect(response.statusCode).toEqual(400);
-    });
-  });
-
-  describe('POST /session/bind-social', () => {
-    it('throw if session is not authorized', async () => {
-      interactionDetails.mockResolvedValueOnce({});
-      await expect(
-        sessionRequest.post('/session/bind-social').send({ connectorId: 'connectorId' })
-      ).resolves.toHaveProperty('statusCode', 400);
-    });
-    it('throw if no social info in session', async () => {
-      interactionDetails.mockResolvedValueOnce({
-        result: { login: { accountId: 'user1' } },
-      });
-      await expect(
-        sessionRequest.post('/session/bind-social').send({ connectorId: 'connectorId' })
-      ).resolves.toHaveProperty('statusCode', 400);
-    });
-    it('updates user identities', async () => {
-      interactionDetails.mockResolvedValueOnce({
-        jti: 'jti',
-        result: {
-          login: { accountId: 'user1' },
-          socialUserInfo: {
-            connectorId: 'connectorId',
-            userInfo: { id: 'connectorUser', phone: 'phone' },
-          },
-        },
-      });
-      const response = await sessionRequest.post('/session/bind-social').send({
-        connectorId: 'connectorId',
-      });
-      expect(response.statusCode).toEqual(200);
-      expect(updateUserById).toHaveBeenCalledWith(
-        expect.anything(),
-        expect.objectContaining({
-          identities: {
-            connectorId: {
-              details: { id: 'connectorUser', phone: 'phone' },
-              userId: 'connectorUser',
-            },
-          },
-        })
-      );
     });
   });
 
@@ -905,4 +293,3 @@ describe('sessionRoutes', () => {
     );
   });
 });
-/* eslint-enable max-lines */
